Fix: satisfy strict mypy in black_scholes (v0.19.0 typecheck gate)
The v0.19.0 bump added a 'mypy --strict' cross-check. scipy ships no type
information, so its calls infer as Any:
- Wrap float-returning expressions in float() to fix no-any-return.
- Add a [tool.mypy] override to ignore scipy.* missing stubs.

+++ src/greeks/black_scholes.py
@@ -15,12 +15,12 @@ class OptionType(StrEnum):
 
 def _d1(S: float, K: float, T: float, r: float, sigma: float) -> float:
     """Compute d1 of the Black-Scholes formula."""
-    return (np.log(S / K) + (r + 0.5 * sigma**2) * T) / (sigma * np.sqrt(T))
+    return float((np.log(S / K) + (r + 0.5 * sigma**2) * T) / (sigma * np.sqrt(T)))
 
 
 def _d2(S: float, K: float, T: float, r: float, sigma: float) -> float:
     """Compute d2 of the Black-Scholes formula (d1 minus sigma*sqrt(T))."""
-    return _d1(S, K, T, r, sigma) - sigma * np.sqrt(T)
+    return float(_d1(S, K, T, r, sigma) - sigma * np.sqrt(T))
 
 
 def price(
@@ -54,9 +54,9 @@ def price(
     d2 = _d2(S, K, T, r, sigma)
     discount = np.exp(-r * T)
     if option_type == OptionType.CALL:
-        return S * norm.cdf(d1) - K * discount * norm.cdf(d2)
+        return float(S * norm.cdf(d1) - K * discount * norm.cdf(d2))
     else:
-        return K * discount * norm.cdf(-d2) - S * norm.cdf(-d1)
+        return float(K * discount * norm.cdf(-d2) - S * norm.cdf(-d1))
 
 
 def delta(
@@ -70,15 +70,15 @@ def delta(
     """First derivative of price with respect to spot."""
     d1 = _d1(S, K, T, r, sigma)
     if option_type == OptionType.CALL:
-        return norm.cdf(d1)
+        return float(norm.cdf(d1))
     else:
-        return norm.cdf(d1) - 1
+        return float(norm.cdf(d1) - 1)
 
 
 def gamma(S: float, K: float, T: float, r: float, sigma: float) -> float:
     """Second derivative of price with respect to spot (same for calls and puts)."""
     d1 = _d1(S, K, T, r, sigma)
-    return norm.pdf(d1) / (S * sigma * np.sqrt(T))
+    return float(norm.pdf(d1) / (S * sigma * np.sqrt(T)))
 
 
 def vega(S: float, K: float, T: float, r: float, sigma: float) -> float:
@@ -87,7 +87,7 @@ def vega(S: float, K: float, T: float, r: float, sigma: float) -> float:
     Returns vega per 1-point move in volatility (not per percentage point).
     """
     d1 = _d1(S, K, T, r, sigma)
-    return S * norm.pdf(d1) * np.sqrt(T)
+    return float(S * norm.pdf(d1) * np.sqrt(T))
 
 
 def theta(
@@ -107,9 +107,9 @@ def theta(
     discount = np.exp(-r * T)
     decay = -(S * norm.pdf(d1) * sigma) / (2 * np.sqrt(T))
     if option_type == OptionType.CALL:
-        return (decay - r * K * discount * norm.cdf(d2)) / 365
+        return float((decay - r * K * discount * norm.cdf(d2)) / 365)
     else:
-        return (decay + r * K * discount * norm.cdf(-d2)) / 365
+        return float((decay + r * K * discount * norm.cdf(-d2)) / 365)
 
 
 def rho(
@@ -127,6 +127,6 @@ def rho(
     d2 = _d2(S, K, T, r, sigma)
     discount = np.exp(-r * T)
     if option_type == OptionType.CALL:
-        return K * T * discount * norm.cdf(d2)
+        return float(K * T * discount * norm.cdf(d2))
     else:
-        return -K * T * discount * norm.cdf(-d2)
+        return float(-K * T * discount * norm.cdf(-d2))
